Increase load addresses to fix boot
With the default load addresses starting at 0x1000000, something seems
to clobber both the FDT and the kernel somewhere early in booti. I'm not
sure what's doing it (doesn't seem to be any Amlogic FDT manipulation
code), but changing the load to start at 0x8000000, like in pyamlboot's
runKernel.py script, fixes it.